Added a condition to not standardize the control bait
Control bait signals are usually very low, and standardizing them for the control values to equal the other baits would not make sense, so the control bait is excluded from the standardization algorithm.

Also, to handle cases of negative values in log2fc(), negative values are now changed to 0 within the function.

Added batch processing support as well.