Optional Kolmogorov-Smirnov screening and conditional matrix column substitution
Uses Kolmogorov-Smirnov testing of a conditional matrix column against a baseline unconditional matrix column - if it passes, it is left alone, but if it fails, it is substituted with the baseline column. This is because if they aren't different, the baseline column is usually less noisy (and thus better) because it is trained on the full dataset rather than a subset.